feat: PB entrega

+++ src/main/java/gabrielborel/com/br/deliveryapp/repositories/SellerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SellerRepository extends CrudRepository<Seller, Long> {
+public interface SellerRepository extends CrudRepository<Seller, Integer> {
 }
